Update README.md
The Modern Literal Version (MLV) is the only bible that wants your input and help to make sure there is NO errors. We have kept this in an "on demand" printing service and Open Source for that reason.  It is sad there are 600 year old printing press typos in the KJV that until it was corrected 37,000 times as the ASV they are still there today.  

The MLV in book form:  
https://www.amazon.com/Modern-Literal-Version-New-Testament-ebook/dp/1492776270/ref=sr_1_1?ie=UTF8&qid=1475284970&sr=8-1&keywords=modern+literal+version

Free PDF's & HTML version from here or www.ModernLiteralVersion.org  (mlvbible.org, .com .net .info)

E-Sword modules are at: 
http://www.biblesupport.com/e-sword-downloads/file/10090-modern-literal-version-2016-greek/  
http://www.biblesupport.com/e-sword-downloads/file/9919-modern-literal-version-2016-beta-edition/

To God is the glory forever & ever.